Use single quotes for JSX props

+++ index.js
@@ -174,7 +174,7 @@ module.exports = {
  */
     'react/display-name': 0,
     'react/jsx-boolean-value': 2,
-    'react/jsx-quotes': [2, 'double'],
+    'react/jsx-quotes': [2, 'single'],
     'react/jsx-no-undef': 2,
     'react/jsx-sort-props': 0,
     'react/jsx-sort-prop-types': 0,
